+++ src/lib/pages/public-project/slug.tsx
@@ -1,12 +1,13 @@
 import { Tabs, TabList, TabPanels, TabPanel } from "@chakra-ui/react";
 import { useRouter } from "next/router";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 
+import { useInternalNavigate } from "lib/app-provider";
 import { CustomTab } from "lib/components/CustomTab";
 import { Loading } from "lib/components/Loading";
 import PageContainer from "lib/components/PageContainer";
 import { AmpEvent, AmpTrack } from "lib/services/amplitude";
-import { scrollToTop } from "lib/utils";
+import { getFirstQueryParam } from "lib/utils";
 
 import { DetailHeader } from "./components/DetailHeader";
 import { PublicProjectAccountTable } from "./components/table/account/PublicProjectAccountTable";
@@ -15,15 +16,16 @@ import { PublicProjectContractTable } from "./components/table/contract/PublicPr
 import { usePublicData } from "./data";
 
 enum TabIndex {
-  Overview,
-  Codes,
-  Contracts,
-  Accounts,
+  Overview = "overview",
+  Codes = "codes",
+  Contracts = "contracts",
+  Accounts = "accounts",
 }
 
-export const ProjectDetail = () => {
+const ProjectDetail = () => {
   const router = useRouter();
-  const [tabIndex, setTabIndex] = useState(TabIndex.Overview);
+  const navigate = useInternalNavigate();
+  const tab = getFirstQueryParam(router.query.tab) as TabIndex;
   const {
     publicCodes,
     publicContracts,
@@ -33,20 +35,44 @@ export const ProjectDetail = () => {
     isLoading,
   } = usePublicData();
 
-  useEffect(() => {
-    if (router.isReady) AmpTrack(AmpEvent.TO_PROJECT_DETAIL);
-  }, [router.isReady]);
-
-  const handleOnViewMore = (tab: TabIndex) => {
-    setTabIndex(tab);
-    scrollToTop();
+  const handleTabChange = (nextTab: TabIndex) => () => {
+    if (nextTab === tab) return;
+    navigate({
+      pathname: "/public-project/[slug]/[tab]",
+      query: {
+        slug,
+        tab: nextTab,
+      },
+      options: {
+        shallow: true,
+      },
+    });
   };
 
+  useEffect(() => {
+    if (router.isReady) {
+      if (!tab || !Object.values(TabIndex).includes(tab)) {
+        navigate({
+          pathname: "/public-project/[slug]/[tab]",
+          query: {
+            slug,
+            tab: TabIndex.Overview,
+          },
+          options: {
+            shallow: true,
+          },
+        });
+      }
+      AmpTrack(AmpEvent.TO_PROJECT_DETAIL, { ...(tab && { tab }) });
+    }
+  }, [router.isReady, tab, slug, navigate]);
+
   if (isLoading) return <Loading />;
+
   return (
     <PageContainer>
       <DetailHeader details={projectDetail} slug={slug} />
-      <Tabs index={tabIndex}>
+      <Tabs index={Object.values(TabIndex).indexOf(tab)}>
         <TabList my={6} borderBottom="1px" borderColor="pebble.800">
           <CustomTab
             count={
@@ -54,28 +80,28 @@ export const ProjectDetail = () => {
               publicContracts.length +
               publicAccounts.length
             }
-            onClick={() => setTabIndex(TabIndex.Overview)}
+            onClick={handleTabChange(TabIndex.Overview)}
           >
             Overview
           </CustomTab>
           <CustomTab
             count={publicCodes.length}
             isDisabled={!publicCodes.length}
-            onClick={() => setTabIndex(TabIndex.Codes)}
+            onClick={handleTabChange(TabIndex.Codes)}
           >
             Codes
           </CustomTab>
           <CustomTab
             count={publicContracts.length}
             isDisabled={!publicContracts.length}
-            onClick={() => setTabIndex(TabIndex.Contracts)}
+            onClick={handleTabChange(TabIndex.Contracts)}
           >
             Contracts
           </CustomTab>
           <CustomTab
             count={publicAccounts.length}
             isDisabled={!publicAccounts.length}
-            onClick={() => setTabIndex(TabIndex.Accounts)}
+            onClick={handleTabChange(TabIndex.Accounts)}
           >
             Accounts
           </CustomTab>
@@ -85,15 +111,15 @@ export const ProjectDetail = () => {
           <TabPanel p={0}>
             <PublicProjectCodeTable
               codes={publicCodes}
-              onViewMore={() => handleOnViewMore(TabIndex.Codes)}
+              onViewMore={handleTabChange(TabIndex.Codes)}
             />
             <PublicProjectContractTable
               contracts={publicContracts}
-              onViewMore={() => handleOnViewMore(TabIndex.Contracts)}
+              onViewMore={handleTabChange(TabIndex.Contracts)}
             />
             <PublicProjectAccountTable
               accounts={publicAccounts}
-              onViewMore={() => handleOnViewMore(TabIndex.Accounts)}
+              onViewMore={handleTabChange(TabIndex.Accounts)}
             />
           </TabPanel>
           <TabPanel p={0}>
@@ -110,3 +136,5 @@ export const ProjectDetail = () => {
     </PageContainer>
   );
 };
+
+export default ProjectDetail;